+++ changelog.md
@@ -11,3 +11,5 @@ date: DD/MM/YYYY
 * Added more methods and EasyMod keys. 13 total!
 # VU5 - 12/01/2022 (19:01) | BonesYT
 * Added 3 more array methods and fixed the .min.js thing.
+# VU6 - 16/01/2022 (22:22) | BonesYT
+* Added approx. 74 functions and 661 lines! Modified and added much more functions and also added more to EasyObj!